requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langgraph==0.2.45
langchain-groq==0.2.1
chromadb==0.5.18
tavily-python==0.5.0
sentence-transformers==3.3.1
streamlit==1.40.0
python-dotenv